+++ R/defaults.R
@@ -24,7 +24,7 @@ default_flextable_settings <- list(
   nan_str  = "",
   fmt_date = "%Y-%m-%d", fmt_datetime = "%Y-%m-%d %H:%M:%S",
 
-  shadow = TRUE, extra_css = "",
+  extra_css = "",
   scroll = NULL,
   split = TRUE, keep_with_next = FALSE,
   tabcolsep = 0, arraystretch = 1.5, float = "none",
@@ -82,9 +82,6 @@ flextable_global$defaults <- default_flextable_settings
 #' xelatex or lualatex. If pdflatex is used, fonts will be ignored because they are
 #' not supported by pdflatex, whereas with the xelatex and lualatex engines they are.
 #' @param extra_css css instructions to be integrated with the table.
-#' @param shadow `TRUE` or `FALSE`, use shadow dom (for HTML only), this option is existing
-#' to disable shadow dom (set to `FALSE`) for pagedown and Quarto that can
-#' not support it for now.
 #' @param scroll NULL or a list if you want to add a scroll-box.
 #' See **scroll** element of argument `opts_html` in function [set_table_properties()].
 #' @param split Word option 'Allow row to break across pages' can be
@@ -109,6 +106,7 @@ flextable_global$defaults <- default_flextable_settings
 #' @param post_process_pdf,post_process_docx,post_process_html,post_process_pptx Post-processing functions
 #' that will allow you to customize the display by output type (pdf, html, docx, pptx).
 #' They are executed just before printing the table.
+#' @param ... unused or deprecated arguments
 #' @return a list containing previous default values.
 #' @examples
 #' ft_1 <- qflextable(head(airquality))
@@ -138,7 +136,6 @@ set_flextable_defaults <- function(
   na_str = NULL, nan_str = NULL,
   fmt_date = NULL, fmt_datetime = NULL,
   extra_css = NULL,
-  shadow = NULL,
   scroll = NULL,
   split = NULL, keep_with_next = NULL,
   tabcolsep = NULL, arraystretch = NULL, float = NULL,
@@ -147,7 +144,8 @@ set_flextable_defaults <- function(
   post_process_pdf = NULL,
   post_process_docx = NULL,
   post_process_html = NULL,
-  post_process_pptx = NULL
+  post_process_pptx = NULL,
+  ...
   ){
 
   x <- list()
@@ -226,9 +224,6 @@ set_flextable_defaults <- function(
   }
   if( !is.null(fonts_ignore) ){
     x$fonts_ignore <- fonts_ignore
-  }
-  if( !is.null(shadow) ){
-    x$shadow <- shadow
   }
   if( !is.null(split) ){
     x$split <- split
@@ -337,7 +332,6 @@ print.flextable_defaults <- function(x, ...){
   if(is.character(x$theme_fun)) cat("## default theme is:", x$theme_fun, "\n")
 
   cat("## HTML specific:\n")
-  cat("shadow:", x$shadow, "\n")
   cat("extra_css:", x$extra_css, "\n")
   cat("scrool:", if (is.null(x$scrool)) "no" else "yes", "\n")
   cat("post_process_html:\n")